Splits isMember into a test and a block, and a block with a loading check
So, now you can check for group membership, or you can check for group membership and *require* it, or you can do that but with also checking for whether the module is being loaded.

For convenience.

Also, isMember let modules load before, but the person couldn't use the software, whereas now it fails to load the module.