Adds the isCpusetLimited proc
This is intended to detect whether the cpuset the current process is
running in does not include all cores.

This is needed as part of a mechanism to work around a problem in Intel
MPI, where certain cpusets can make it segfault if a certain
configuration value is not set to make it use the secondary topology
determination mechanism instead of the default one.